Refactor UI components by extracting variants and hooks into separate files
