+++ src/components/ui/button-variants.ts
@@ -1,0 +1,34 @@
+import { cva } from "class-variance-authority";
+
+const buttonVariants = cva(
+  "inline-flex items-center justify-center gap-2 whitespace-nowrap rounded-lg text-sm font-medium ring-offset-background transition-all duration-200 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 [&_svg]:pointer-events-none [&_svg]:size-4 [&_svg]:shrink-0 hover:transform active:scale-95",
+  {
+    variants: {
+      variant: {
+        default: "bg-primary text-primary-foreground hover:bg-primary/90 hover:scale-105 shadow-soft hover:shadow-hover",
+        destructive:
+          "bg-destructive text-destructive-foreground hover:bg-destructive/90 hover:scale-105",
+        outline:
+          "border border-input bg-background hover:bg-accent hover:text-accent-foreground hover:scale-105 hover:shadow-md",
+        secondary:
+          "bg-secondary text-secondary-foreground hover:bg-secondary/80 hover:scale-105",
+        ghost: "hover:bg-accent hover:text-accent-foreground hover:scale-105",
+        link: "text-primary underline-offset-4 hover:underline hover:scale-105",
+        hero: "bg-gradient-hero text-primary-foreground hover:scale-110 shadow-soft hover:shadow-hover hover:shadow-primary/20 font-semibold",
+        glass: "bg-card/80 backdrop-blur-sm border border-border/50 text-card-foreground hover:bg-card/90 hover:scale-105 shadow-soft hover:shadow-hover",
+      },
+      size: {
+        default: "h-10 px-4 py-2",
+        sm: "h-9 rounded-md px-3",
+        lg: "h-11 rounded-md px-8",
+        icon: "h-10 w-10",
+      },
+    },
+    defaultVariants: {
+      variant: "default",
+      size: "default",
+    },
+  }
+);
+
+export { buttonVariants };
